requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Diffusion Model Libraries
diffusers>=0.25.0
accelerate>=0.25.0
transformers>=4.36.0

# Data Processing
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# Explainability
shap>=0.43.0
captum>=0.6.0
lime>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
tensorboard>=2.14.0
wandb>=0.16.0

# Evaluation Metrics
pytorch-fid>=0.3.0
clean-fid>=0.1.35
scipy>=1.10.0

# Configuration Management
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.66.0
einops>=0.7.0
timm>=0.9.0
ema-pytorch>=0.3.0

# Logging and Monitoring
loguru>=0.7.0
rich>=13.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Optional: For advanced features
# einops-exts>=0.0.4
# rotary-embedding-torch>=0.3.0
# x-transformers>=1.27.0

